for ui Events and score Events, switch from String + parsec to ByteString + attoparsec
This is dubious because non-ascii characters are now silently corrupted (not
even UTF8 is supported, though it could be) and the performance gain isn't as
much as I thought it would be.  However, symbols should be represented with the
symbol notation anyway, and it still feels like bytestrings should be much more
efficient, in space at least.

In any case, if it turns out to be unwise, it is easy enough to revert.