make derivation lazy
Signals are still strict, so this is only part of the job.

- tracks and events are run in a separate Deriver monad, to keep the ErrorT
strictness contained

- Derivers now return the new LEvent type, which integrates the logs into the
output stream, so the calling function has control over the amount of
evaluation.  Deriver is still in LogT so logs must be merged into the output
when sub-derives are run.  LEvent logs are preserved all the way through
conversion and performance.  This breaks tons of tests since 'derive' output is
no longer Either.

- Add Convert_test and clean up tests while I'm at it.  Lazy_test is intended
to verify laziness but it's incomplete.