add cmds to instrument definitions
This wound up changing instrument definitions quite a bit because MidiDb cannot
import Cmd.Cmd due to circular imports, and it was awkward how the calls and
cmds would have to be in separate modules from the instruments themselves.

I removed all of the code to resolve the ModuleIds to the actual code at
startup in favor of instruments directly containing or importing the code for
their calls and cmds.  This seems simpler and lets instruments be more
self-contained.  Since the code is still not serializable I have to strip it
off before serialization and restore it on unserialization, which means each
instrument must store its own cache instead of having one big one.

- main app no longer loads search index, since only the browser uses it
- search index is now generated on load, rather than being cached to disk
- move instrument definition utilities to a new App.MidiInst module