fix the generator cache so it interacts properly with inverting calls
This involved some major restructuring of the generator cache.

For the cache to work with inverting calls, it needs to be checked below the
inversion, at the actual generating call.  Otherwise it can't notice damage on
the control track.  This means the cache check has to be a normal part of a
call like Note.inverting_call rather than a special field of GeneratorCall
handled at the top level.  It had to be at the top level because it needed to
be called even for non-caching calls, to record ControlDamage.  So I switched
to an alternate scheme where ControlDamage is extrapolated directly from the
ScoreDamage and the event extends with no reference to the generated signal.

This should be more efficient as well since now ControlDamage is calculated
only once per track rather than on every single call, but in the profiles it
makes very little difference.

Now that neither ControlDamage nor EventDamage use local damage, that field
can be removed entirely.  This means my scheme for the transform cache won't
work, but maybe the transform cache will never be needed or if it is I can use
the same technique as ControlDamage.