replace transformer stack DeriveT with custom monad
It uses continuations instead of ErrorT, and it was supposed to be faster
because at one point the profiler showed a significant amount of allocation
was Eithers.  However this is apparently no longer the case because this
change makes no noticeable difference.

Oh well, might as well keep it, now that I have it.