add orphan track extraction
This extracts notes that are beneath another note track but not covered by
one of its events, and thus would not otherwise be evaluated.  This is
especially convenient for a track to apply ornaments to several tracks
below it, and when there is no ornament to be applied the tracks below are
still played.

There's still no convenient way to apply an ornament to varying sets of notes,
but I'll think up some sort of solution for that if it becomes a problem.