fix problem where cmds that invalidate a lot of score cause a hang
The problem was that Cmd.Track.get_track_cmds wants to find out the
instrument to get instrument cmds, and that in turn demands
Derive.TrackEnviron, which evidently demands quite a bit of derivation.

I should try to make TrackEnviron lazier, but the more complete solution was
to eliminate the ability for cmds to trigger a lot of derivation
synchronously by not replacing the performance until the evaluation thread
has forced some important parts of it.  This means that play will continue
to play the old performance which is a bit annoying, but I think it's
worth the tradeoff.