fix bug where modifying a tempo track above a control track would make derivation fail
The problem was that the real to score for the control damage of the tempo
track was done outside of the warp established by the tempo track itself,
which meant it would generate damage ranges that were incorrect for the tracks
below it, which are inside the warp.

Calculating time relative to the correct warp is tricky.