change signal functions so that the value before the first sample is taken to be zero
Previously, an empty signal was taken to be zero, and a signal with one sample
was a constant of that value, no matter where the sample actually was.  This
seems counter-intuitive to me so I changed it.

This fixes a Perform bug where a pedal down would set the signal but since
there was no previous 0 then the pedal was considered to have always been
down.

Updated Perform.Signal doc and moved it to Perform.SignalBase while I was at
it.