fix bug where a note under a note transformer was incorrectly cached after score damage
This actually turned out to be fairly involved but was a bad bug.  Namely,
slicing moves events, so they are no longer in track time, but the cache, via
Stack.Region and ControlDamage assumes that the ScoreTimes it gets are in
track time.  Also it makes more sense to the user.  The fix was to pass the
sliced track range via tevents_range in TrackEvents, and the lower end of the
range is how much the events were shifted back.  So adding it back to the
ScoreTime gets back into track time.

This is fiddly business and it I tried to add a separate TrackTime type, but
it turned out to be a hassle to convert back and forth, and hopefully I can
get this right in the core functions and the various calls won't have to
worry about it.

Since I calculate ControlDamage from events rather than from generated signal,
it turns out it's no longer necessary to keep it in RealTime, which means I
can remove that unpleasant score -> real -> score conversion.