fix bug where the controls of sliced events weren't trimmed enough
They were using block_end as the next event, which is incorrect if there
is a next event, but it's been sliced off.  I replaced block_end with
event_end, which is more specifically meant to be the end of the event's
range, whether that be the start of the next event or the end of the block.