remove the schema concept entirely
This was a per-block mapping to a block deriver.  It used to also map to
the set of track-specific cmds.  The original idea was that the kind of
notation and cmd set could be entirely reconfigured per-block by setting
a custom SchemaId, but in practice the non-track specific set of cmds came
to depend on the default schema so eventually cmds were moved out.  Now that
block calls are normal calls and there is a Scopes concept that can
dynamically rebind symbols, I can theoretically configure how blocks are
derived via those means.  This does mean I lose the ability to set the root
block deriver and to completely reconfigure cmds, but I can add those features
if I ever need them.

In practice, I'm not sure I'll ever want a completely different kind of
notation, since it would by definition be incompatible with all the existing
notation.  It seems better to try to figure out how to incorporate any new
notation concept into the existing structure.

The parser from Derive.Schema was split into Derive.ParseSkeleton, and the
block derivers into Derive.Call.BlockUtil.