implement chord mode
This involved a fair amount of refactoring.  Instead of Cmd.Track figuring out
track type information and passing it to the various cmds, they figure out the
types on their own and the type detection functions move to Cmd.Info.
Cmd.Track should be dealing with track types beyond the minimum necessary to
decide which cmds apply.