let inverted calls find out the controls of following notes
This is so things like the rambat tick can work properly.

The solution was to modify inversion slicing so it can pick up extra events
from the control tracks, and then put yet another field into State.TrackEvents
so that the inverted call can know about the previous and next events even
though they are outside of its slice.  I don't like that a lot, but I guess
it's generally useful for inverted calls to have access to all the same score
information that non-inverted ones do, especially if it winds up that most
everything is inverted.

It's not super pretty but it's a lot nicer than the other things I was
considering, as documented in the TODO.