history is not recorded for edit cmds while edit mode is still active
This has the effect of grouping everything done when edit mode is on into one
undo transaction.  I don't know if I'll definitely want this, but it will cut
down on the amount of history to save, and it's analogous to how vim groups
all edits in edit mode into one undo.