incremental save: remove events from Update.TrackUpdate
This means updates are no longer sufficient to transform a State, you need
the destination State since events (and CreateView's view) are no longer in
the updates.  This means Ui.ApplyUpdate can't work anymore, so it gets
deleted.  I had intended to use it for incremental saves, but using git means
I don't have to deal with diffs myself.  Meanwhile, the presence of events
in TrackUpdates meant that diffs had directionality since they recorded
the specific events in the destination event and couldn't then be used for
undoing into that state from the future.  This made undo more complicated.

It seems a shame to use ApplyUpdate but I never did wind up using it.