fix bug where Cmd.Perf.get_scale_id on a note track doesn't pick up the scale on a track underneath it
To fix this I had to give the synthesized inverted track the TrackId of its
source track, so it shows up in the stack and can record the environ from the
post-invert perspective, at which point it will know things like the scale of
pitch tracks it was inverted under.  This means every inverted note will
redundantly record an environ and be overwritten, but hopefully that's
harmless.

It also changes what the stack looks like, but hopefully this is harmless
too, and is actualy a more accurate picture of the flow of evaluation.