integrate: fix bug where diff couldn't handle notes with multiple control events
The problem is that the stack is inadequate since it's the stack of the
single score event that emitted all of the control events.  Since I'm
only diffing within tracks now I don't really need the stack anyway.

So I simplified the index key to just be the event's original position on the
track.  This is not perfect, but better than the stack, and hopefully easy to
understand when it's wrong.