major refactor: put start pos in Ui.Event.Event and remove Ui.Events.PosEvent
It's a relatively minor change but affects many signatures in many places.

While I'm at it, restrict exports from Ui.Event and Ui.Events, and reorganize
serialization a bit due to that.

I'm not 100% sure this is a great idea, but it certainly seems simpler to
deal with only Event instead of Event and PosEvent, which was just a
(ScoreTime, Event) pair.  In exchange I lose the ability to refer to events
without positinos, e.g. (Event -> Event) was not able to change the start
time.  It also means the start time is duplicated in the track map 'Map
ScoreTime Event', but that's concealed inside Ui.Events so hopefully they can
never get out of sync.