generalize Score.event_attributes to event_environ
This is because the lilypond performer wants more context for notes, like
left or right hand, clef, time signature, etc.  It means I lose Read for
score events, since the Environ has Vals, and some of them have code, but I
should be able to get that back if it's important.