the deriver considers blocks to end at the end of the ruler, instead of at the last event
This is more convenient since you don't have to add '--' events on a block
with trailing rests, and corresponds to the Cmd layer, which uses the ruler.

The change was very simple but wound up breaking a lot of tests in fiddly
ways because they relied on the block length.

Along the way I fixed a bug that caused note trills to go to the next event
instead of the end of the current one.