include call documentation in the calls themselves so it can be extracted with introspection
This adds a bit of complexity to call declarations, but in return I get
online documentation for all calls, which seems like a big win especially as
there gets to be more and more.

It wound up needing a bunch more ShowVal instances, so ShowVal needed its own
module to avoid circular imports.