fix bug where manually replacing the ruler made the ruler track small
Turns out it's due to a long sequence of events: replacing the ruler deletes
and re-adds all tracks since that's the only way to change the ruler.  It
deletes from 0, so each track gets a turn in the ruler slot.  But
set_ruler_width didn't reinitialize its internal sizes, so it only worked
the first time.  So the ruler track size gets stuck in the wrong place.

It might be nice to delete tracks from the top down, e.g. 2, 1, 0 instead of
0, 0, 0, but it might make Seq.indexed_pairs a little hacky, and in any case
this bug needed to be fixed.