make ornaments that create events before time 0 work
This happens when a note at 0 has a grace note.  To make the negative events
have defined controls, I put a hack in TimeVector.at that says a sample at <=0
extends backwards as well as forwards, and Cmd.Play tweaks the start time to
account for the negative events.

I'm not too happy about changing the signal definition yet again but at least
I have a lengthy rationale in place for the next time.