change control track samples from a custom parser to a special lookup
This way is cleaner since I don't have to have a separate parser for control
tracks.  That in turn makes it difficult to write a generic function, like
Derive.Call.eval_event, that parses events because it would have to need to
know what kind of track it's on.

I previously didn't do it this way because I was afraid of the performance
implications, but according to profiling it's the same.