try to control lag and void by strictifying derivation
It helps, but there are still significant amounts of void and drag.
Interestingly, there's almost no lag.

This marks a change in strategy from "lazy as possible" derivation to
explicitly evaluating it all before putting it in the performances.  Due to
signals being strict, especially the tempo signal, and fltk signal rendering,
it was basically strict anyway, but in a way that left a lot of unevaluated
drag.

The new idea is that I should forget about derive-level laziness (performance
is still lazy) and instead try to make it as fast as possible.  If I'm
committed to strictness I can also try to parallelize.