remove the last traces of the pitch signal ValName support
This let pitch signals display scale degree references.  But eventually I
decided it's not worth it, because it has to be absolute to make sense as
a graph, but then it can't easily represent scales whose degrees are not
absolute.  What's more, it didn't actually look that good since it turns out
to be more useful to see the pitch signal normalized from lowest to highest
pitch rather than to each scale degree.
can't easily represent scales whose