lilypond: make grace notes inside a tuplet work
This involved interleaving lilypond "performance" with derivation, so
derivers can embed little chunks of lilypond code.  I also wound up having
to stick the lilypond config into Derive.Deriver.Constant instead of the
Environ.  It's really more appropriate there because it's constant and typed.
It means a bit of special lilypond knowledge is embedded in the deriver,
which I wanted to avoid, but oh well.