make +staccato shorten the notes in addition to applying an attr
I also changed legato to simply add the attribute, and left it up to
the note deriver to interpret +legato and +staccato.  This is more consistent
with allowing the note call to make the decisions, and will let an instrument
change the interpretation of legato or staccato by swapping out the note
call.