lilypond: go back to having the lilypond performer interpret attributes
This is more in keeping with letting the lower levels interpret higher level
instructions.  The benefit is that the lilypond will get the articulation no
matter how the attribute was added.  This is especially handy for things
like +pizz, which don't really need a special call when the attribute
notation is already concise enough.