fiddle key bindings yet again, hopefully for the last time
I forgot that using just Fl::event_key() doesn't work because then I can't
enter any shifted chars.  So I now include the key and the text separately,
the key is used to map actions to physical keys, and the text is used to enter
text.

Hopefully this is the last time I'll need to mess with key bindings.