refine how note slicing works a bit
Instead of only slicing to the events of the next note track down and looking
for orphans, I look all the way to the bottom of the branch and slice to all
the events, where higher events cover up lower ones.  This is about the same
as before, except that all slices are merged into one list per track, instead
of orphans looking like their own tracks as before.  This makes calls work
where they didn't before.

I also throw an error when a sub-event has two parents, since that never
worked right, and in fact can't since the call graph has to be a tree, not a
graph.

More details in blog post titled 'slicing'.