lilypond: add implementation for separate movements
There are actually two implementations.  The first adds a 'movement' call
which splits the score into multiple movemenst.  It's nice because the
movement boundaries and titles are in the score, but not so nice because
you have to create a top-level block for the score, and as usual with
lilypond derivation, it ignores the tempo track.  So you have to create
a giant score block whose timing has to be 1:1 with the entire piece.

The second lets you derive multiple blocks at once, and provide them with
movement titles.  It's nice because it's simpler than the first
implementation and doesn't force a whole-score block, but not so nice
because the movement titles are outside of the score, on the REPL command
line.

I'm not sure which one I want, so I'll keep them both until I figure that
out.