remove the fltk implementation of signal labels
This was a feature that would let a signal add labels to various values,
and the signal render would then annotate the signal when it crossed those
values.  It was originally to label pitch signals with scale degrees, but
that went by the wayside when pitch signals became abstract.  Since the
frequency of scale degrees can now depend on the environment, it's
impossible for the fltk layer to do the reverse transformation and label a
given NN.