make TimeSteps search from the given point
Previously they worked by building a list of all the time steps on the
block, and then stepping through them.  With laziness only the points
before the step position need to be evaluated, but it's still linear in
the number of steps.  Even though I haven't been able to pin any
performance problems on TimeStep, it's always made me unhappy because
there can be a lot of time steps and the snap function is called every
time the cursor moves on a drag.  I don't have any proof this is faster,
but it sure seems like it should be.