stash only TrackSignals explicitly asked for
Previously I saved TrackSignals for all tracks and let laziness do its
thing.  Unfortunately, this also creates thunks which hold a reference to
all of the old derivation state, which causes a space leak.  This way is
less elegant but is more efficient with space.

A better solution would be to figure out why holding the reference to the
old state would cause space to leak endlessly, rather than just for one
derivation cycle.