add support for relative intervals to just scales
Since the relative intervals can be based on the previous pitch, this
required more robust support for ValCalls to get the previous pitch value.
I wound up adding a new Tagged type which goes in 'Derive.info_prev_val'
for ValCalls.  This is more reliable than the previous technique of
just evaluating the previous event because it works even when the previous
event itself relies on its previous event.