add *twelve-r scale, which is relative notation for 12TET
This turned out to be a lot more involved than I thought, and it's still
not quite right because the absolute keyboard interacts poorly with
relative scales when accidentals enter the picture.

The twelve scale is now closer to how the just scale works, since it has to
rely on parsing instead of a degree map.