change Perform.Pitch.InputKey to Input
InputKey was basically just a NoteNumber, and it was up to each scale to
figure out what to do with that.  But this turned out to not be enough,
since I wanted to treat the ASCII kbd differently from the piano kbd.
So Input now has a lot more information, namely a KbdType, along with a
Theory.Pitch with octave, pitch class, and accidentals.

This generalization lets me treat the ASCII kbd as a 10-key keyboard with
an accidental between each diatonic key.

It would up being way more complicated and epic than I expected.