split generator and transformer calls
Previously both generator and transformer were in the same Call data type.
But this meant that if, say, an instrument override a generator it would
also replace a transformer with the same name.  I originally thought that
it would be less confusing since many generators and transformers are
paired, but now I think it's just an unnecessary restriction.

However, there are some calls that do have an analogous generator and
transformer that want to be defined together, I added a few abstractions to
deal with them as pairs.  Fundamentally, however, they now exist in
entirely separate namespaces.

This involved touching every single call so it was a huge hassle.
Hopefully I won't have to do much more of this.