add three more Scale methods: scale_read, scale_show, and scale_layout
With these I can write a generic to_relative, as well as convert between
scales (e.g. absolute to relative).  I'm reluctant to add three methods
just for this, but they seem like features I want and this seems to be the
simplest way to do it.  Perhaps I can now get rid of scale_transpose, but
I'm not sure.