shakefile: instead of having .hi files generated from .hs.o files by doing nothing, make .hs emit both .o and .hi
I think it should wind up the same, but it seems like shake is better at
reporting circular dependencies in this case.

Also add the buildHs dep on hsconfig.h back.  It turns out it's necessary
because otherwise the include-chaser can't find it and doesn't know what
directory it's supposed to be in.  Later I can fix the include-chaser to
have a special case for hsconfig.h.