allow equal expressions without spaces, e.g. a=b
It seems nice to be able to minimize space usage in the score.  Also it
makes them not wrap, which can look nicer.  The catch is that I can't use
'=' in call names anymore.  I don't have any at the moment, so it doesn't
seem like to heavy a price to pay.