make velocity come from control functions so it can be randomized
This was surprisingly complicated, since it interacts with the Pressure
flag.  I initially tried to get rid of Pressure entirely and put it all in
the note call, but would up leaving it in Perform.Midi.Convert and
communicating the control function dynamic via a magic control,
Controls.dynamic_function.

As a side effect, NoteOff velocities are now always the same as NoteOn,
since velocity is stored as a scalar, not a signal, which broke all the
regression tests.  I don't think it's a big deal since I don't care about
NoteOff velocity yet.