generalize c_equal to bind quoted expressions
This creates a new call that is effectively a macro for the bound
expression.  Currently quoted expressions are syntactically limited to val
calls, but I plan on lifting that soon.

This is moving in the direction of user-defined functions, which I'm
nervous about.  What happend to trying to do all this stuff in haskell?