drop the strict naming enforcement in Ui.Id
It was pretty complicated and a lot of callers wound up using the unsafe
functions anyway.  In any case, having unapproved characters in an ID
mostly just means you can't call it from a tracklang literal, which is
hardly a disaster.  It still seems good to enforce consistency in naming,
so as long as you go through State.read_id you'll still get an error on bad
characters.