get rid of the val call fallback
This was originally a mechanism to allow scale degrees to be used as both
val calls and as pitch calls.  But it seemed simpler to have the scale
explicitly put val calls and pitch calls in scope.

The old behaviour meant you could also directly use non-scale val calls for
pitch and control calls, e.g. you could write `<` but must now write
`set (<)`.  This is convenient, but perhaps not worth complicating the
call evaluation code.  If I really wanted to get it back, I could also
implement it as a lookup that looks in the val call scope, rather than
hardcoding in Derive.Call.apply_generator.