switch event text and tracklang parser from ByteString to Text
I originally used UTF8-encoded ByteStrings because attoparsec didn't support
text, and attoparsec-text, which did, was slower in profiles.  Now it appears
that switching to Text improves performance, so I can finally get rid of
the awkward ByteStrings.

Best of all, no saved scores are broken since the serialization for UTF8
ByteString and Text is the same!