rewrite Derive.Call.Bali.Reyong to use Kotekan functions
Now it behaves similarly to the kotekan calls, which is that you write
calls that emit individual melody notes, instead of writing a melody and
having it infer the kilitan that should come out of that.  It's a bit lower
level, but is consistent with how kotekan as done as well as being much
simpler.

In addition, since I now have the Scale.scale_read and Scale.scale_show
functions, it works for any scale.