rewrite most of Bali.Reyong yet again, to support kotekan patterns
This time I use a table of absolute pitches to patterns, along with a
lot of functions to generate those tables, as opposed to the mixture of
tables and code that I had before.  Hopefully this will be able to account
for pattern variations and irregularities that will come later.