rework how orphan event slicing works
This is the result of an epic yak shaving expedition that started with
trying to fix TrackDynamic recording for blocks with multiple callers.
Those changes are still in, though not complete yet.  But that in turn
revealed that titles of sliced orphans weren't being evaluated when their
tracks were stripped due to being empty, which in turn meant their
TrackDynamics wouldn't be recorded correctly.  That in turn made me
dissatisfied with the complexity of the existing mechanism.

In the end, I changed orphans from special code in BlockUtil.derive_tracks
that separately extracted the orphans and evaluated them out of order, to
extra code in 'Call.derive_track' (specialized to 'derive_note_track') to
call 'Slice.slice_notes' on empty stretches.  This has the same effect but
gets rid of all of the separate orphan slicing code in 'Derive.Slice' since
it is now merged with 'Slice.slice_notes'.

It also turned up a number of bugs in note slicing that all basically
reduced to me not fully understanding the 'TrackTree.tevents_range' and
'tevents_end' fields.  I still don't, but I plan to clean those up soon.