complete work to simplify TrackDynamic recording
Now I reccord TrackDynamics from non-inverted and inverted tracks, and
combine them at the end, instead of relying on the order of track
evaluation and Monoid hacks.

A side effect of this change is that I now combine Collects by mappending
the new one to the right, instead of the left.  In cases where it matters,
as for Data.Map, this means that the first occurrence of some key (e.g.
(BlockId, TrackId) for TrackDynamic) will win, not the last.