fix bug in new orphan slicing scheme that prevented calls from getting previous control values
I wound up rejiggering orphan slicing yet again, in that I'm back to a
normal slice, rather than a note slice.  This way seems simpler and more
efficient, I'm just going to evaluate the orphan slice (which could be the
entire track, if the parent is empty!) in sequence, and don't need to split
it into individual notes.

This wound up turning up yet more problems with overlap detection, and I
still think it's broken, but the tests pass so I'm giving up on it for now.