redo the prev val mechanism to explicitly save and restore the previous val
Previously I relied on slicing to include extra control events, and when
that wasn't enough, evaluate the previous control call on demand.  Now I
save the previous value in a new Threaded state.  This adds a linear data
dependency on derivation, which is crummy, but in practice it existed
already in the prev value, now it's just explicit.  It can still be broken
at block boundaries.

Originally this was to add prev val support for note calls, which I also
did, but it's actually kind of orthogonal to the main prev val change.
There are also numerous refactors and cleanups.

Another somewhat orthogonal change is that the Derive.Elem type family is
gone, and I save the entire previous signal or event.  I could reinstate
this, but it seems simpler this way.  It may be less efficient because
I store the entire previous chunk just to extract the last value, but it
likely doesn't make a difference since most control calls emit a single
chunk, so mconcat has no work to do.