collect TrackSignals only from top level blocks
The previous behavour was inefficient, because it collected TrackSignals
redundantly, including for blocks with no view.  In addition, it tended to
use the TrackSignals from the root derive, which are much more likely to
require an unwarp.  It also was inconsistent, ResponderSync used the root
derive, while the DeriveComplete response used whatever block just
completed.

Only collecting TrackSignals for the top level blocks should solve all of
those problems.  Also, I now derive all visible blocks, not just the one
with focus.