change track titles to use WrappedInput instead of ExpandInput
ExpandInput wasn't so great because it would frequently run out of
horizontal space on a track close to the right edge of the window.
Wrapping vertically seems to be both easier to read and to edit, but does
mean that a narrow track will wrap a lot.

That was the last use of ExpandInput so I can get rid of it entirely.  That
means getting rid of all the icky hacks necessary when ExpandInput
overlapped its neighbors.