allow a selnum to address multiple discontiguous selections
This actually simplifies things a bit.  Since fltk now has a lower level
idea of what a selection is, I could get rid of set_track_selection and the
TrackSelection type since all fltk level selections are now per-track.

Now a selnum is just a awy to address an arbitrary set of selections.