add standalone ngoret calls
To get them to work they needed previous and next events, which made me wind
up implementing a mechanism to "delay" calls until their appropriate
postproc call.  It seems like overkill, but it seems like needing to know
neighbor notes is a general problem that will come up again in the future.

Ngoret now requires an accompanying realize-ngoret call as a global
transform, and wants 'hand' annotations, but in return it can now infer
pitches and handle damping across block boundaries.  So it's a bit less
convenient, but more correct.