add Environ.track_voice, and Derive.Call.InferTrackVoice
This keeps track of the count of tracks with the same instrument, counting
from the left.  Also add c_infer_hands, which uses that to set the first
track to hand=l and the second to hand=r.

This is a way to infer score structure from the placement of the tracks.
I'm not sure I really like it because it's implicit, but on the other hand
I have implicit things to avoid cluttering the score with repetetive stuff
like hand assignments.