change Score.unwarp_pos to always return a ScoreTime instead of a Maybe
Previously I would fail if the RealTime is out of the range of the Warp,
with the assumption that asking for an out-of-range time is probably a
mistake.  However, ngoret grace notes wind up doing this due to overlap
time, so it's time to revisit this.

Now I implicitly extend the Warp at the same tempo as its start and end.
I previously extended Score.warp_pos in the same way, so now unwarp_pos is
consistently the inverse of unwarp_pos.