change how notes get control signals
Instead of the default note deriver trimming each signal to end at the start
of the next event or the end of the block, the note deriver now does not
trim the end of the signal at all, and relies on the performer to trim it
according to the instrument's decay.

This is because notes from a block call weren't seeing control changes from
after the end of the block.  Of course, the note has already ended at that
point, but it can still matter for notes with a long decay, or a pedal
control, which artificially lengthens the note duration.

I'm not totally sure I like this change, since the previous behaviour was
convenient, and meant I didn't really have to care if the inst_decay was
correct.  But otherwise it's confusing how a block can't see controls
outside of its extent.