remove split control feature
I've never used it, and it's a pain to keep its test passing.

For posterity, this was a features that let you put an event like '%c' in
a control track.  It would have the effect of changing the track's control
signal to '%c' until the next time a '%' event occurred.