change how inversion re-invokes the inverted call
Previously, I saved the expression in Derive.info_expr, and inversion
created a fake track and event, and the expression was then re-interpreted.
A call tag and a hack in Eval caused transformers be skipped on the second
interpretation, except ones with a special Tags.under_invert.

The problem with this scheme was that I had to be careful to update
info_expr when I evaluated an expression, or a later inversion would wind
up re-interpreting something different.  I had a number of hacks to deal
with this in different circumstances, but it remained error-prone and hard
to even know what the right thing is.

So the new approach is that I have two new fields in Derive.Dynamic,
state_under_invert and state_inversion.  These save the inverted call
directly in Dynamic, and track evaluation directly applies the saved
deriver when it notices that it's deriving a synthesized inverted track.
state_under_invert does the same for under_invert transformers.

This means I can get rid of the whole info_expr mess.  Also it should be a
bit more efficient, since I don't have to interpret and typecheck every
inverting expression twice.