make it so control msgs extend until the next note, not until the end of the decay
The problem is that the pedal control changes note durations, which means I
wind up clipping off control changes that would have been audible, such as
pedal up.  A real solution would be to make pedal extend note ends in the
performer, but meanwhile this works and also makes it once again not so
critical to get note decay time right.