sort events after derivation
Previously I relied on the events being in order, and tried to keep them
that way by doing sorted merges.  However, this was pretty error-prone
since any function that emitted unsorted events would break this property.
For instance, any postproc that moves an event is likely to break sorting.

I decided that sorting all the events after derivation is fast enough
(perhaps even faster, since I don't have to work to maintain the order),
and is less complicated and error-prone.  Ultimately this is another
consequence of no longer trying to make derivation lazy.