prevent apply-start-offset from moving events past neighbors on the same hand
I also guess if the note is supposed to line up to its successor, and
if it is, change the duration to maintain that.

Originally this was a bug fix where events would overlap and get cut off
due to MIDI not being able to express the same pitch overlapping itself,
but turned into something much more elaborate when I considered what should
happen when two events wanted to move into the same place.  It seems like
it should be straightforward, but for whatever reason turned out anything
but.