make the performer not trim controls if there are no next notes
It was cutting off pedal ups.  These are problematic because they extend
the duration of notes, but the performer doesn't know about this.  A
principled fix would probably be to teach the performer about the special
nature of the pedal control, but for the moment just not trimming controls
seems to work.