add interpolate scale
This involved a surprising amount of refactoring and redesign of how
pitches work.

Firstly, I have to parameterize the scale with other scales, which means it
needs to get an Environ for configuration at creation time.  So scales
are now either Scale.Simple, which is just a Scale, and Scale.Make, which
expects arguments to create the scale.

The bigger problem was that since pitches got the Environ at conversion
time, the interpolate scale couldn't substitute `key-from` and `key-to`
because any environ values it set before evaluating the pitch wouldn't
"stick".

So I switched to applying the environ when the pitch is created, but this
broke the case that made me switch to applying environ at the end in the
first place, which is when switching instruments in a postproc I want the
new instrument's environ, specifically Environ.tuning, to apply to the new
event's pitches.  So now when I switch instruments I have to explicitly
apply the environ from the new instrument, via Post.set_instrument.

Along the way, it seemed clearer to stash the PitchConfig in the Pitch as a
separate field instead of just composing it onto the pitch_nn and pitch_note
functions.