make TrackDynamic get scale from the inverted track, but everything else from the non-inverted track
Originally I got everything from the inverted track except the controls, but
I don't want note entry to get confused when an instrument generates events
with different instruments, e.g. >kontakt/kendang-pasang.  But I still want
the scale, because it's common to set the scale in the inverted track.

Actually, maybe it's not that common, usually all instruments are in a
single global scale.