fix a bug where rederives didn't get the right damage
The problem was that the damage was taken from the current performance,
while the actual derivation proceeded from the evaluated performance.  So
if there was sequential damage, the first bit of damage would be forgotten
since it would be immediately cleared, even though its derivation was
cancelled by the second edit.

The solution was to put damage in both Cmd.state_performance and
Cmd.state_current_performance.