refactor aliases and PlayUtil
Originally I meant to reuse the instrument alias map from StateConfig in
the Derive, to make the error msg clearer when you alias to an instrument
that has no allocation.  But it turns out to break things because Derive
actually has to alias to the StateConfig alias, which has the allocation,
rather than the source instrument, which just supplies the configuration.